Theres something wrong with imagecreatefromstring

+++ merge.php
@@ -8,7 +8,8 @@ if ($_SESSION["login"] == "guest") {
     index_error(-1, "You must be logged in to view this page");
 }
 $pdo = DB::getConnection();
-if (isset($_POST["images"])) {
+if (isset($_POST["images"]))
+{
     $all = json_decode($_POST["images"]);
 
     $full_thing = imagecreatetruecolor(500, 375);
@@ -17,55 +18,64 @@ if (isset($_POST["images"])) {
     $fw = imagesx($full_thing);
     $fh = imagesy($full_thing);
     $org_aspect = $fw / $fh;
-
-    foreach ($all as $key => $value) {
+	$fail = false;
+    foreach ($all as $key => $value)
+	{
 
         $data = explode(",", $value);
-        $data[1] = base64_decode($data[1]);
-        if(($img = imagecreatefromstring($data[1])))
+		$fresh = base64_decode($data[1]);
+
+		$img = imagecreatefromstring($fresh);
+        if($img !== false)
 	   	{
 			imagealphablending($img, true);
 			imagesavealpha($img, true);
 			$w = imagesx($img);
 			$h = imagesy($img);
 			if ($w / $h > $org_aspect)
-			{
 				$w = $h*$org_aspect;
-			}
 			else
 				$h = $w/$org_aspect;
-			//echo $w." ".$h.PHP_EOL;
 			$img = imagescale($img, $fw, -1);
 			imagecopy($full_thing, $img, 0, 0, 0, 0, $w, imagesy($img));
 		}
+		else
+			$fail = true;
     }
-    try {
-        $imgdir = "./imgs/trial" . $_SESSION["user_id"] . ".png";
-        imagepng($full_thing, $imgdir);
-        $final_image = base64_encode(file_get_contents($imgdir));
-        if (isset($_GET)) {
-            if ($_GET["type"] == "propic") {
-                $stmt = $pdo->prepare("UPDATE users SET `avatar`=:img, `type`='png'WHERE id=:uid");
-                $stmt->bindParam(":img", $final_image, PDO::PARAM_STR);
-                $stmt->bindParam(":uid", $_SESSION["user_id"]);
-                $stmt->execute();
-            } else {
+    try{
+			if ($fail == false)
+			{
+				$imgdir = "./imgs/trial" . $_SESSION["user_id"] . ".png";
+				imagepng($full_thing, $imgdir);
+				$final_image = base64_encode(file_get_contents($imgdir));
+				if (isset($_GET)) {
+					if ($_GET["type"] == "propic") {
+						$stmt = $pdo->prepare("UPDATE users SET `avatar`=:img, `type`='png'WHERE id=:uid");
+						$stmt->bindParam(":img", $final_image, PDO::PARAM_STR);
+						$stmt->bindParam(":uid", $_SESSION["user_id"]);
+						$stmt->execute();
+					} else {
 
-                $stmt = $pdo->prepare("INSERT INTO `images` (`user_id`, `src`, `creation_date`, `type`) VALUES (:uid, :src, NOW(), 'png')");
-                $stmt->bindParam(":uid", $_SESSION["user_id"], PDO::PARAM_INT);
-                $stmt->bindParam(":src", $final_image, PDO::PARAM_STR);
-                $stmt->execute();
-            }
-            imagedestroy($full_thing);
-            unlink($imgdir);
-            echo "success";
-            exit();
-        }
-        echo "failure";
-    } catch (\PDOException $e) {
+						$stmt = $pdo->prepare("INSERT INTO `images` (`user_id`, `src`, `creation_date`, `type`) VALUES (:uid, :src, NOW(), 'png')");
+						$stmt->bindParam(":uid", $_SESSION["user_id"], PDO::PARAM_INT);
+						$stmt->bindParam(":src", $final_image, PDO::PARAM_STR);
+						$stmt->execute();
+					}
+					imagedestroy($full_thing);
+					unlink($imgdir);
+					echo "success";
+					exit();
+				}
+			}
+			echo "failure";
+		}
+        
+    catch (\PDOException $e) {
         echo $e->getMessage();
     }
-} else {
+}
+else
+{
     echo "Failure: Invalid Method/File";
 }
